backend/requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic
openai
python-dotenv
pytest
httpx
